requirements.txt
Pillow==10.0.0
Pillow==10.0.0
PyMuPDF==1.22.5
pytesseract==0.3.10
